update assignment controller and other changes

+++ backend/controllers/assignmentController.js
@@ -6,8 +6,8 @@ const mongoose = require("mongoose");
 
 const createAssignment = (request, response) => {
   const { name, title, due, description, username, creator } = request.body;
-  console.log("Giving jwt", request.jwtObj)
-  //if (request.jwtObj) {
+  console.log("Giving jwt", request.jwtObj);
+  if (request.jwtObj) {
     Parent.findOne({ username: username })
       .then(user => {
         const assignment = new Assignment({
@@ -21,28 +21,39 @@ const createAssignment = (request, response) => {
         assignment
           .save()
           .then(saveAssignment => {
-            response.status(200).json(saveAssignment);
+            const id = saveAssignment._id;
+            Parent.findOneAndUpdate(
+              { username: username },
+              { $push: { assignments: id } }
+            ).then(saveAssignment => {
+              response.status(200).json(saveAssignment);
+            });
+          })
+          .catch(err => {
+            response
+              .status(500)
+              .json({ errorMessage: "Error saving assignment", err });
           })
           .catch(err => {
             console.log("Error here", err);
             response
               .status(500)
-              .json({ message: "Error creating assignment", err });
+              .json({ errorMessage: "Error creating assignment", err });
           });
       })
       .catch(err => {
         console.log("Error here", err);
         response
           .status(500)
-          .json({ message: "Error creating assignment", err });
+          .json({ errorMessage: "Error creating assignment", err });
       });
-  // } else {
-  //   response.status(422).json({ message: "User Not Logged In" });
-  // }
+  } else {
+    response.status(422).json({ errorMessage: "User Not Logged In" });
+  }
 };
 const getAssignmentsByParent = (request, response) => {
   const { username } = request.params;
-  //if (request.jwtObj) {
+  if (request.jwtObj) { //authenticate user
     Parent.findOne({ username: username })
       .then(user => {
         id = user._id;
@@ -57,11 +68,11 @@ const getAssignmentsByParent = (request, response) => {
       .catch(err => {
         response.status(500).json(err);
       });
-  // } else {
-  //   return response
-  //     .status(422)
-  //     .json({ error: "Login is required before assignments can be viewed" });
-  // }
+  } else {
+    return response
+      .status(422)
+      .json({ errorMessage: "Login is required before assignments can be viewed" });
+  }
 };
 const getAllAssignments = (request, response) => {
   Assignment.find({})
@@ -74,42 +85,45 @@ const getAllAssignments = (request, response) => {
 };
 const deleteAssignment = (request, response) => {
   const { _id } = request.body;
-  // if (request.jwtObj) {  
-   Assignment.findOne({_id: request.params._id})
-   .then(assignment => {
-     const id = assignment.creator;
-     //Here we delete referenced id of deleted assignment from Parent
-     Parent.findOneAndUpdate({_id: id}, { $pull: {assignments:  request.params._id}})
-     .then(user => {      
-       Assignment.findOneAndRemove({ _id: request.params._id })
-       .then(assignment => {
-         response.status(200).json(assignment);
-       })
-       .catch(err => {
-         aresponse.status(500).json({
-           errorMessage: "Something went wrong while deleting assignment",
-           err
-         });
-       })
-       .catch(err => {
-         response.status(500).json({
-           errorMessage: "Something went wrong while finding assignment",
-           err
-         });
-       })   
-     })
-       })
-     .catch(err => {
-       response.status(500).json({
-         errorMessage: "Something went wrong while deleting assignment from assignments array",
-         err
-       });
-     });
-     // } else {
-  //   return response
-  //     .status(422)
-  //     .json({ error: "Login is required before activity can be viewed" });
-  // }
+  if (request.jwtObj) {
+    Assignment.findOne({ _id: request.params._id })
+      .then(assignment => {
+        const id = assignment.creator;
+        //Here we delete referenced id of deleted assignment from Parent
+        Parent.findOneAndUpdate(
+          { _id: id },
+          { $pull: { assignments: request.params._id } }
+        ).then(user => {
+          Assignment.findOneAndRemove({ _id: request.params._id })
+            .then(assignment => {
+              response.status(200).json(assignment);
+            })
+            .catch(err => {
+              aresponse.status(500).json({
+                errorMessage: "Something went wrong while deleting assignment",
+                err
+              });
+            })
+            .catch(err => {
+              response.status(500).json({
+                errorMessage: "Something went wrong while finding assignment",
+                err
+              });
+            });
+        });
+      })
+      .catch(err => {
+        response.status(500).json({
+          errorMessage:
+            "Something went wrong while deleting assignment from assignments array",
+          err
+        });
+      });
+  } else {
+    return response
+      .status(422)
+      .json({ errorMessage: "Login is required before activity can be viewed" });
+  }
 };
 const updateAssignment = (request, response) => {
   const { _id, user, title, due, description } = request.body;
